refactoring file types into a package
