+++ rfi_file_monitor/file.py
@@ -2,7 +2,7 @@ import gi
 
 gi.require_version("Gtk", "3.0")
 from gi.repository import Gtk, GLib, Gio
-from .utils import match_path
+
 
 from enum import auto, IntEnum, unique
 import logging
@@ -193,183 +193,8 @@ class File(ABC):
         GLib.idle_add(self._update_progressbar_worker_cb, index, value)
 
 
-class RegularFile(File):
-    def __init__(
-        self,
-        filename: str,
-        relative_filename: PurePath,
-        created: int,
-        status: FileStatus,
-    ):
-
-        super().__init__(
-            filename,
-            relative_filename,
-            created,
-            status,
-        )
 
 
-class WeightedRegularFile(RegularFile):
-    def __init__(
-        self,
-        filename: str,
-        relative_filename: PurePath,
-        created: int,
-        status: FileStatus,
-        offset: float,  # fractional
-        weight: float,  # fractional
-    ):
-
-        super().__init__(
-            filename,
-            relative_filename,
-            created,
-            status,
-        )
-
-        self._offset = offset
-        self._weight = weight
-
-    def update_progressbar(self, index: int, value: float):
-        new_value = 100.0 * self._offset + value * self._weight
-        GLib.idle_add(self._update_progressbar_worker_cb, index, new_value)
 
 
-class Directory(File):
-    def __init__(
-        self,
-        filename: str,
-        relative_filename: PurePath,
-        created: int,
-        status: FileStatus,
-        included_patterns: List[str],
-        excluded_patterns: List[str],
-    ):
 
-        super().__init__(
-            filename,
-            relative_filename,
-            created,
-            status,
-        )
-
-        self._included_patterns = included_patterns
-        self._excluded_patterns = excluded_patterns
-
-        self._filelist: List[Tuple[str, int]] = []
-        self._filelist_timestamp: int = 0
-        self._total_size: int = 0
-
-    @property
-    def included_patterns(self):
-        return self._included_patterns
-
-    @property
-    def excluded_patterns(self):
-        return self._excluded_patterns
-
-    def _get_filelist(self, _dir: Path) -> List[Tuple[str, int]]:
-        rv: List[Tuple[str, int]] = []
-        for entry in _dir.iterdir():
-            if not match_path(
-                entry,
-                included_patterns=self._included_patterns,
-                excluded_patterns=self._excluded_patterns,
-                case_sensitive=False,
-            ):
-                continue
-            if entry.is_file() and not entry.is_symlink():
-                size = entry.stat().st_size
-                self._total_size += size
-                rv.append(
-                    (
-                        str(entry),
-                        size,
-                    )
-                )
-            elif entry.is_dir() and not entry.is_symlink():
-                rv.extend(self._get_filelist(entry))
-        return rv
-
-    def _refresh_filelist(self):
-        self._total_size = 0
-        self._filelist = self._get_filelist(Path(self.filename))
-        self._filelist_timestamp = time()
-
-    @property
-    def total_size(self):
-        if self._filelist_timestamp < self._saved:
-            self._refresh_filelist()
-        return self._total_size
-
-    def __iter__(self):
-        if self._filelist_timestamp < self._saved:
-            self._refresh_filelist()
-        yield from self._filelist
-
-    def __len__(self):
-        return len(self._filelist)
-
-
-class URL(File):
-    def __init__(
-        self,
-        filename: str,
-        relative_filename: PurePath,
-        created: int,
-        status: FileStatus,
-    ):
-
-        super().__init__(
-            filename,
-            relative_filename,
-            created,
-            status,
-        )
-
-
-class S3Object(File):
-    def __init__(
-        self,
-        filename: str,
-        relative_filename: PurePath,
-        created: int,
-        status: FileStatus,
-        bucket_name: str,
-        etag: str,
-        size: int,
-        region_name: str = "",
-    ):
-
-        super().__init__(
-            filename,
-            relative_filename,
-            created,
-            status,
-        )
-        self._bucket_name = bucket_name
-        self._etag = etag
-        self._size = size
-        self._key = str(PurePosixPath(*self._relative_filename.parts))
-        self._region_name = region_name
-
-    @property
-    def bucket_name(self):
-        return self._bucket_name
-
-    @property
-    def etag(self):
-        return self._etag
-
-    @property
-    def key(self):
-        return self._key
-
-    @property
-    def size(self):
-        return self._size
-
-    @property
-    def region_name(self):
-        return self._region_name
