+++ setup.cfg
@@ -63,7 +63,10 @@ rfi_file_monitor.engines =
 rfi_file_monitor.preferences =
     AllowedFilePatternsPreference = rfi_file_monitor.preferences:AllowedFilePatternsPreference
     IgnoredFilePatternsPreference = rfi_file_monitor.preferences:IgnoredFilePatternsPreference
-gui_scripts =
+rfi_file_monitor.files =
+    RegularFile = rfi_file_monitor.files:RegularFile
+    WeightedRegualrFile = rfi_file_monitor.files:WeightedRegularFile
+console_scripts =
     rfi-file-monitor = rfi_file_monitor:main
 
 [bdist_wheel]
